refactor analysis
