+++ cocoa/turk/accept_negotiation_hits.py
@@ -1,12 +1,13 @@
-__author__ = 'anushabala'
-from boto.mturk.connection import MTurkConnection
-from boto.mturk.price import Price
-from boto.mturk.connection import MTurkRequestError
-from argparse import ArgumentParser
 import csv
 import sqlite3
 import json
-from cocoa.analysis.negotiation.utils import get_total_tokens_per_agent, get_turns_per_agent
+from argparse import ArgumentParser
+from boto.mturk.connection import MTurkConnection
+from boto.mturk.price import Price
+from boto.mturk.connection import MTurkRequestError
+
+#TODO
+from analysis.utils import get_total_tokens_per_agent, get_turns_per_agent
 
 MIN_TOKENS = 40
