+++ negotiation/analysis/predict.py
